Fikset review kommentar

+++ src/main/java/no/nav/foreldrepenger/autotest/klienter/fptilbake/behandlinger/dto/BrukerresponsDto.java
@@ -1,12 +1,12 @@
 package no.nav.foreldrepenger.autotest.klienter.fptilbake.behandlinger.dto;
 
 public class BrukerresponsDto {
-    protected int behandlingId;
+    protected Long behandlingId;
     protected boolean akseptertFaktagrunnlag;
     protected String kildeKanal;
 
     public BrukerresponsDto(int behandlingId) {
-        this.behandlingId = behandlingId;
+        this.behandlingId = (long) behandlingId;
         this.kildeKanal = "MANUELL";
     }
